+++ ext/ruby_prof/rp_measure_wall_time.c
@@ -17,7 +17,7 @@ static double measure_wall_time(rb_trace_arg_t* trace_arg)
 #if defined(_WIN32)
     LARGE_INTEGER time;
     QueryPerformanceCounter(&time);
-    return time.QuadPart;
+    return (double)time.QuadPart;
 #elif defined(__APPLE__)
     return mach_absolute_time();// * (uint64_t)mach_timebase.numer / (uint64_t)mach_timebase.denom;
 #elif defined(__linux__)
